Notification 엔티티 수정, ChatService 작성 완료, PaymentDto 결제 금액 필드 타입 변경, SubscriptionRequestDto에서 UserType 필드 삭제, NotificationService 미구현, PaymentService 미구현

+++ backend/src/main/java/com/jobPrize/admin01_service/dto/PaymentResponseDto.java
@@ -17,6 +17,6 @@ public class PaymentResponseDto {
 	private Long id;	// 사용자 ID
 	private UserType userType;	// 사용자 구분
 	private Long paymentId;	// 주문 번호
-    private Long paymentAmount;	// 결제 금액
+    private int paymentAmount;	// 결제 금액
     private LocalDateTime createdTime;  // 결제 시각
 }
